Quick Fix to display the zero with a dash
I implemented this quick fix to show the dash with the roboto mono font. It's not a perfect solution because more letters than the 0 are changed with the font but it's better then before. Doesn't look too easy to change just one letter in css without maybe swrewing up the layout